requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Caching==2.1.0
python-dotenv==1.0.0
# pyodbc==5.0.1  # Removed - SQL Server driver not available on Linux/Render
sqlalchemy[sqlite]==3.1.1
requests==2.32.5
WTForms==3.1.1
email-validator==2.1.1
bleach==6.1.0
qrcode==8.0
Pillow==10.4.0
APScheduler==3.10.4
MarkupSafe==2.1.1
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
werkzeug==3.0.0
cryptography==41.0.7
gunicorn==22.0.0
